Add Notes field to Inventory Items
When creating a new Item in the inventory there is an option to add
notes.
These notes Item carry over as the default entry to the PackItem notes
when added to a Pack.
To save space on the Inventory list, a blue dot at the start
of the line item shows that notes are attached. A grey dot
shows that no notes have been written on the item. When
hovering over the dot, the notes are shown if present.